refactor(tree): moved inherent methods into AllocState trait impls
