+++ bsan-rt/src/tree_borrows/tree.rs
@@ -25,7 +25,7 @@ use super::diagnostics::{
 use super::foreign_access_skipping::IdempotentForeignAccess;
 use super::perms::{AccessKind, PermTransition, Permission};
 use super::tree_visitor::{ChildrenVisitMode, ContinueTraversal, NodeAppArgs, TreeVisitor};
-use super::wildcard::ExposedCache;
+use super::wildcard::{ExposedCache, WildcardAccessLevel};
 use crate::errors::UBResult;
 use crate::helpers::{AllocRange, FxHashSet, Size};
 use crate::sanitizer_common::Span;
@@ -387,94 +387,6 @@ impl EagerTree {
 }
 
 impl EagerTree {
-    /// Insert a new tag in the tree.
-    ///
-    /// `inside_perm` defines the initial permissions for a block of memory starting at
-    /// `base_offset`. These may nor may not be already marked as "accessed".
-    /// `outside_perm` defines the initial permission for the rest of the allocation.
-    /// These are definitely not "accessed".
-    fn new_child(
-        &mut self,
-        base_offset: Size,
-        parent_tag: BorTag,
-        new_tag: BorTag,
-        inside_perms: DedupRangeMap<LocationState>,
-        outside_perm: Permission,
-        protected: bool,
-        span: Span,
-    ) -> UBResult<()> {
-        let idx = self.tag_mapping.insert(new_tag);
-        let parent_idx = if parent_tag.is_wildcard() {
-            None
-        } else {
-            Some(self.tag_mapping.get(&parent_tag).unwrap())
-        };
-        assert!(outside_perm.is_initial());
-
-        let default_strongest_idempotent =
-            outside_perm.strongest_idempotent_foreign_access(protected);
-        // Create the node
-        self.nodes.insert(
-            idx,
-            Node {
-                tag: new_tag,
-                parent: parent_idx,
-                children: SmallVec::default(),
-                default_initial_perm: outside_perm,
-                default_initial_idempotent_foreign_access: default_strongest_idempotent,
-                is_exposed: false,
-                debug_info: NodeDebugInfo::new(new_tag, outside_perm, span),
-            },
-        );
-        if let Some(parent_idx) = parent_idx {
-            let parent_node = self.nodes.get_mut(parent_idx).unwrap();
-            // Register new_tag as a child of parent_tag
-            parent_node.children.push(idx);
-        } else {
-            // If the parent had wildcard provenance, then register the idx
-            // as a new wildcard root.
-            // This preserves the orderedness of `roots` because a newly created
-            // tag is greater than all previous tags.
-            self.roots.push(idx);
-        }
-
-        // We need to know the weakest SIFA for `update_idempotent_foreign_access_after_retag`.
-        let mut min_sifa = default_strongest_idempotent;
-        for (Range { start, end }, &perm) in
-            inside_perms.iter(Size::from_bytes(0), inside_perms.size())
-        {
-            assert!(perm.permission.is_initial());
-            assert_eq!(
-                perm.idempotent_foreign_access,
-                perm.permission.strongest_idempotent_foreign_access(protected)
-            );
-
-            min_sifa = cmp::min(min_sifa, perm.idempotent_foreign_access);
-            for (_range, loc) in self
-                .locations
-                .iter_mut(Size::from_bytes(start) + base_offset, Size::from_bytes(end - start))
-            {
-                loc.perms.insert(idx, perm);
-            }
-        }
-
-        // We don't have to update `exposed_cache` as the new node is not exposed and
-        // has no children so the default counts of 0 are correct.
-
-        // If the parent is a wildcard pointer, then it doesn't track SIFA and doesn't need to be updated.
-        if let Some(parent_idx) = parent_idx {
-            // Inserting the new perms might have broken the SIFA invariant (see
-            // `foreign_access_skipping.rs`) if the SIFA we inserted is weaker than that of some parent.
-            // We now weaken the recorded SIFA for our parents, until the invariant is restored. We
-            // could weaken them all to `None`, but it is more efficient to compute the SIFA for the new
-            // permission statically, and use that. For this we need the *minimum* SIFA (`None` needs
-            // more fixup than `Write`).
-            self.update_idempotent_foreign_access_after_retag(parent_idx, min_sifa);
-        }
-
-        Ok(())
-    }
-
     /// Restores the SIFA "children are stronger"/"parents are weaker" invariant after a retag:
     /// reduce the SIFA of `current` and its parents to be no stronger than `strongest_allowed`.
     /// See `foreign_access_skipping.rs` and [`Tree::new_child`].
@@ -520,226 +432,11 @@ impl EagerTree {
         }
     }
 
-    /// Deallocation requires
-    /// - a pointer that permits write accesses
-    /// - the absence of Strong Protectors anywhere in the allocation
-    fn dealloc(
-        &mut self,
-        tag: BorTag,
-        access_range: AllocRange,
-        global: &GlobalState,
-        alloc_id: AllocId, // diagnostics
-        span: Span,        // diagnostics
-    ) -> UBResult<()> {
-        self.perform_access(
-            tag,
-            access_range,
-            AccessKind::Write,
-            AccessCause::Dealloc,
-            global,
-            alloc_id,
-            span,
-        )?;
-
-        let start_idx =
-            if tag.is_wildcard() { None } else { Some(self.tag_mapping.get(&tag).unwrap()) };
-
-        // Check if this breaks any strong protector.
-        // (Weak protectors are already handled by `perform_access`.)
-        for (loc_range, loc) in self.locations.iter_mut(access_range.start, access_range.size) {
-            let diagnostics = DiagnosticInfo {
-                alloc_id,
-                span,
-                transition_range: loc_range,
-                access_range: Some(access_range),
-                access_cause: AccessCause::Dealloc,
-            };
-            // Checks the tree containing `idx` for strong protector violations.
-            // It does this in traversal order.
-            let mut check_tree = |idx| {
-                TreeVisitor { nodes: &mut self.nodes, data: loc }
-                    .traverse_this_parents_children_other(
-                        idx,
-                        // Visit all children, skipping none.
-                        |_| ContinueTraversal::Recurse,
-                        |args: NodeAppArgs<'_, _>| {
-                            let node = args.nodes.get(args.idx).unwrap();
-
-                            let perm = args
-                                .data
-                                .perms
-                                .get(args.idx)
-                                .copied()
-                                .unwrap_or_else(|| node.default_location_state());
-                            if global.get_protector_kind(node.tag)
-                                == Some(ProtectorKind::StrongProtector)
-                                // Don't check for protector if it is a Cell (see `unsafe_cell_deallocate` in `interior_mutability.rs`).
-                                // Related to https://github.com/rust-lang/rust/issues/55005.
-                                && !perm.permission.is_cell()
-                                // Only trigger UB if the accessed bit is set, i.e. if the protector is actually protecting this offset. See #4579.
-                                && perm.accessed
-                            {
-                                Err(TbError {
-                                    error_kind: TransitionError::ProtectedDealloc,
-                                    access_info: &diagnostics,
-                                    conflicting_node_info: &node.debug_info,
-                                    accessed_node_info: start_idx
-                                        .map(|idx| &args.nodes.get(idx).unwrap().debug_info),
-                                }
-                                .build())
-                            } else {
-                                Ok(())
-                            }
-                        },
-                    )
-            };
-            // If we have a start index we first check its subtree in traversal order.
-            // This results in us showing the error of the closest node instead of an
-            // arbitrary one.
-            let accessed_root = start_idx.map(&mut check_tree).transpose()?;
-            // Afterwards we check all other trees.
-            // We iterate over the list in reverse order to ensure that we do not visit
-            // a parent before its child.
-            for &root in self.roots.iter().rev() {
-                if Some(root) == accessed_root {
-                    continue;
-                }
-                check_tree(root)?;
-            }
-        }
-        Ok(())
-    }
-
-    /// Map the per-node and per-location `LocationState::perform_access`
-    /// to each location of the first component of `access_range_and_kind`,
-    /// on every tag of the allocation.
-    ///
-    /// `LocationState::perform_access` will take care of raising transition
-    /// errors and updating the `accessed` status of each location,
-    /// this traversal adds to that:
-    /// - inserting into the map locations that do not exist yet,
-    /// - trimming the traversal,
-    /// - recording the history.
-    fn perform_access(
-        &mut self,
-        tag: BorTag,
-        access_range: AllocRange,
-        access_kind: AccessKind,
-        access_cause: AccessCause, // diagnostics
-        global: &GlobalState,
-        alloc_id: AllocId, // diagnostics
-        span: Span,        // diagnostics
-    ) -> UBResult<()> {
-        #[cfg(feature = "expensive-consistency-checks")]
-        if self.roots.len() > 1 || matches!(prov, ProvenanceExtra::Wildcard) {
-            self.verify_wildcard_consistency(global);
-        }
-
-        let source_idx =
-            if tag.is_wildcard() { None } else { Some(self.tag_mapping.get(&tag).unwrap()) };
-
-        // We iterate over affected locations and traverse the tree for each of them.
-        for (loc_range, loc) in self.locations.iter_mut(access_range.start, access_range.size) {
-            let diagnostics = DiagnosticInfo {
-                access_cause,
-                access_range: Some(access_range),
-                alloc_id,
-                span,
-                transition_range: loc_range,
-            };
-            loc.perform_access(
-                self.roots.iter().copied(),
-                &mut self.nodes,
-                source_idx,
-                access_kind,
-                global,
-                ChildrenVisitMode::VisitChildrenOfAccessed,
-                &diagnostics,
-                /* min_exposed_child */ None, // only matters for protector end access,
-            )?;
-        }
-        Ok(())
-    }
-    /// This is the special access that is applied on protector release:
-    /// - the access will be applied only to accessed locations of the allocation,
-    /// - it will not be visible to children,
-    /// - it will be recorded as a `FnExit` diagnostic access
-    /// - and it will be a read except if the location is `Unique`, i.e. has been written to,
-    ///   in which case it will be a write.
-    /// - otherwise identical to `EagerTree::perform_access`
-    fn perform_protector_end_access(
-        &mut self,
-        tag: BorTag,
-        global: &GlobalState,
-        alloc_id: AllocId, // diagnostics
-        span: Span,        // diagnostics
-    ) -> UBResult<()> {
-        #[cfg(feature = "expensive-consistency-checks")]
-        if self.roots.len() > 1 {
-            self.verify_wildcard_consistency(global);
-        }
-
-        let source_idx = self.tag_mapping.get(&tag).unwrap();
-
-        let min_exposed_child = if self.roots.len() > 1 {
-            LocationTree::get_min_exposed_child(source_idx, &self.nodes)
-        } else {
-            // There's no point in computing this when there is just one tree.
-            None
-        };
-
-        // This is a special access through the entire allocation.
-        // It actually only affects `accessed` locations, so we need
-        // to filter on those before initiating the traversal.
-        //
-        // In addition this implicit access should not be visible to children,
-        // thus the use of `traverse_nonchildren`.
-        // See the test case `returned_mut_is_usable` from
-        // `tests/pass/tree_borrows/tree-borrows.rs` for an example of
-        // why this is important.
-        for (loc_range, loc) in self.locations.iter_mut_all() {
-            // Only visit accessed permissions
-            if let Some(p) = loc.perms.get(source_idx)
-                && let Some(access_kind) = p.permission.protector_end_access()
-                && p.accessed
-            {
-                let diagnostics = DiagnosticInfo {
-                    access_cause: AccessCause::FnExit(access_kind),
-                    access_range: None,
-                    alloc_id,
-                    span,
-                    transition_range: loc_range,
-                };
-                loc.perform_access(
-                    self.roots.iter().copied(),
-                    &mut self.nodes,
-                    Some(source_idx),
-                    access_kind,
-                    global,
-                    ChildrenVisitMode::SkipChildrenOfAccessed,
-                    &diagnostics,
-                    min_exposed_child,
-                )?;
-            }
-        }
-        Ok(())
-    }
 }
 
 #[allow(unused)]
 /// Integration with Miri's garbage collector
 impl EagerTree {
-    fn remove_unreachable_tags(&mut self, live_tags: &FxHashSet<BorTag>) {
-        for i in 0..(self.roots.len()) {
-            self.remove_useless_children(self.roots[i], live_tags);
-        }
-        // Right after the GC runs is a good moment to check if we can
-        // merge some adjacent ranges that were made equal by the removal of some
-        // tags (this does not necessarily mean that they have identical internal representations,
-        // see the `PartialEq` impl for `UniValMap`)
-        self.locations.merge_adjacent_thorough();
-    }
-
     /// Checks if a node is useless and should be GC'ed.
     /// A node is useless if it has no children and also the tag is no longer live.
     fn is_useless(&self, idx: UniIndex, live: &FxHashSet<BorTag>) -> bool {
@@ -1220,120 +917,6 @@ impl LazyTree {
         }
     }
 
-    /// Returns true if `tag` is present in this tree.
-    /// In the `Uninit` state only the root tag exists.
-    fn contains_tag(&self, tag: BorTag) -> bool {
-        match self {
-            LazyTree::Uninit { root_tag, .. } => *root_tag == tag,
-            LazyTree::Init(tree) => tree.tag_mapping.contains_key(&tag),
-        }
-    }
-
-    /// Returns the number of nodes in the tree (1 when `Uninit`).
-    fn node_count(&self) -> usize {
-        match self {
-            LazyTree::Uninit { .. } => 1,
-            LazyTree::Init(tree) => tree.tag_mapping.len(),
-        }
-    }
-
-    /// Adds a child tag. Forces initialization of the underlying `EagerTree` first.
-    fn new_child(
-        &mut self,
-        base_offset: Size,
-        parent_tag: BorTag,
-        new_tag: BorTag,
-        inside_perms: DedupRangeMap<LocationState>,
-        outside_perm: Permission,
-        protected: bool,
-        span: Span,
-    ) -> UBResult<()> {
-        self.ensure_init();
-        let LazyTree::Init(tree) = self else { unreachable!() };
-        tree.new_child(
-            base_offset,
-            parent_tag,
-            new_tag,
-            inside_perms,
-            outside_perm,
-            protected,
-            span,
-        )
-    }
-
-    /// Performs an access. In the `Uninit` state (single root node with `Unique` permission,
-    /// no protectors), any access by the root tag is trivially valid (so we short-circuit with `Ok(())`).
-    fn perform_access(
-        &mut self,
-        tag: BorTag,
-        access_range: AllocRange,
-        access_kind: AccessKind,
-        access_cause: AccessCause,
-        global: &GlobalState,
-        alloc_id: AllocId,
-        span: Span,
-    ) -> UBResult<()> {
-        match self {
-            LazyTree::Uninit { .. } => Ok(()),
-            LazyTree::Init(tree) => tree.perform_access(
-                tag,
-                access_range,
-                access_kind,
-                access_cause,
-                global,
-                alloc_id,
-                span,
-            ),
-        }
-    }
-
-    /// Performs a deallocation. In the `Uninit` state the root always has write permission and
-    /// no protectors exist, so this is a no-op.
-    fn dealloc(
-        &mut self,
-        tag: BorTag,
-        access_range: AllocRange,
-        global: &GlobalState,
-        alloc_id: AllocId,
-        span: Span,
-    ) -> UBResult<()> {
-        match self {
-            LazyTree::Uninit { .. } => Ok(()),
-            LazyTree::Init(tree) => tree.dealloc(tag, access_range, global, alloc_id, span),
-        }
-    }
-
-    /// Performs the implicit access on protector release. In the `Uninit` state no protectors
-    /// can exist (they are only registered when children are added), so this is a no-op.
-    fn perform_protector_end_access(
-        &mut self,
-        tag: BorTag,
-        global: &GlobalState,
-        alloc_id: AllocId,
-        span: Span,
-    ) -> UBResult<()> {
-        match self {
-            LazyTree::Uninit { .. } => Ok(()),
-            LazyTree::Init(tree) => tree.perform_protector_end_access(tag, global, alloc_id, span),
-        }
-    }
-
-    /// Marks `tag` as exposed. No-op for `Uninit` trees since the root is always `Unique`
-    /// and its permission can never be invalidated without a second node existing.
-    fn expose_tag(&mut self, tag: BorTag, protected: bool) {
-        self.ensure_init();
-        if let LazyTree::Init(tree) = self {
-            tree.expose_tag(tag, protected);
-        }
-    }
-
-    #[allow(unused)]
-    /// GC pass: no-op on `Uninit` trees since there are no unreachable children.
-    fn remove_unreachable_tags(&mut self, live_tags: &FxHashSet<BorTag>) {
-        if let LazyTree::Init(tree) = self {
-            tree.remove_unreachable_tags(live_tags);
-        }
-    }
 }
 
 /// Relative position of the access
@@ -1404,10 +987,16 @@ pub trait AllocState: Clone {
 
 impl AllocState for LazyTree {
     fn contains_tag(&self, tag: BorTag) -> bool {
-        self.contains_tag(tag)
+        match self {
+            LazyTree::Uninit { root_tag, .. } => *root_tag == tag,
+            LazyTree::Init(tree) => tree.tag_mapping.contains_key(&tag),
+        }
     }
     fn node_count(&self) -> usize {
-        self.node_count()
+        match self {
+            LazyTree::Uninit { .. } => 1,
+            LazyTree::Init(tree) => tree.tag_mapping.len(),
+        }
     }
     fn new_child(
         &mut self,
@@ -1419,15 +1008,9 @@ impl AllocState for LazyTree {
         protected: bool,
         span: Span,
     ) -> UBResult<()> {
-        self.new_child(
-            base_offset,
-            parent_tag,
-            new_tag,
-            inside_perms,
-            outside_perm,
-            protected,
-            span,
-        )
+        self.ensure_init();
+        let LazyTree::Init(tree) = self else { unreachable!() };
+        tree.new_child(base_offset, parent_tag, new_tag, inside_perms, outside_perm, protected, span)
     }
     fn perform_access(
         &mut self,
@@ -1439,7 +1022,12 @@ impl AllocState for LazyTree {
         alloc_id: AllocId,
         span: Span,
     ) -> UBResult<()> {
-        self.perform_access(tag, access_range, access_kind, access_cause, global, alloc_id, span)
+        match self {
+            LazyTree::Uninit { .. } => Ok(()),
+            LazyTree::Init(tree) => {
+                tree.perform_access(tag, access_range, access_kind, access_cause, global, alloc_id, span)
+            }
+        }
     }
     fn dealloc(
         &mut self,
@@ -1449,7 +1037,10 @@ impl AllocState for LazyTree {
         alloc_id: AllocId,
         span: Span,
     ) -> UBResult<()> {
-        self.dealloc(tag, access_range, global, alloc_id, span)
+        match self {
+            LazyTree::Uninit { .. } => Ok(()),
+            LazyTree::Init(tree) => tree.dealloc(tag, access_range, global, alloc_id, span),
+        }
     }
     fn perform_protector_end_access(
         &mut self,
@@ -1458,13 +1049,21 @@ impl AllocState for LazyTree {
         alloc_id: AllocId,
         span: Span,
     ) -> UBResult<()> {
-        self.perform_protector_end_access(tag, global, alloc_id, span)
+        match self {
+            LazyTree::Uninit { .. } => Ok(()),
+            LazyTree::Init(tree) => tree.perform_protector_end_access(tag, global, alloc_id, span),
+        }
     }
     fn expose_tag(&mut self, tag: BorTag, protected: bool) {
-        self.expose_tag(tag, protected)
+        self.ensure_init();
+        if let LazyTree::Init(tree) = self {
+            tree.expose_tag(tag, protected);
+        }
     }
     fn remove_unreachable_tags(&mut self, live_tags: &FxHashSet<BorTag>) {
-        self.remove_unreachable_tags(live_tags)
+        if let LazyTree::Init(tree) = self {
+            tree.remove_unreachable_tags(live_tags);
+        }
     }
 }
 
@@ -1485,15 +1084,59 @@ impl AllocState for EagerTree {
         protected: bool,
         span: Span,
     ) -> UBResult<()> {
-        self.new_child(
-            base_offset,
-            parent_tag,
-            new_tag,
-            inside_perms,
-            outside_perm,
-            protected,
-            span,
-        )
+        let idx = self.tag_mapping.insert(new_tag);
+        let parent_idx = if parent_tag.is_wildcard() {
+            None
+        } else {
+            Some(self.tag_mapping.get(&parent_tag).unwrap())
+        };
+        assert!(outside_perm.is_initial());
+
+        let default_strongest_idempotent =
+            outside_perm.strongest_idempotent_foreign_access(protected);
+        self.nodes.insert(
+            idx,
+            Node {
+                tag: new_tag,
+                parent: parent_idx,
+                children: SmallVec::default(),
+                default_initial_perm: outside_perm,
+                default_initial_idempotent_foreign_access: default_strongest_idempotent,
+                is_exposed: false,
+                debug_info: NodeDebugInfo::new(new_tag, outside_perm, span),
+            },
+        );
+        if let Some(parent_idx) = parent_idx {
+            let parent_node = self.nodes.get_mut(parent_idx).unwrap();
+            parent_node.children.push(idx);
+        } else {
+            self.roots.push(idx);
+        }
+
+        let mut min_sifa = default_strongest_idempotent;
+        for (Range { start, end }, &perm) in
+            inside_perms.iter(Size::from_bytes(0), inside_perms.size())
+        {
+            assert!(perm.permission.is_initial());
+            assert_eq!(
+                perm.idempotent_foreign_access,
+                perm.permission.strongest_idempotent_foreign_access(protected)
+            );
+
+            min_sifa = cmp::min(min_sifa, perm.idempotent_foreign_access);
+            for (_range, loc) in self
+                .locations
+                .iter_mut(Size::from_bytes(start) + base_offset, Size::from_bytes(end - start))
+            {
+                loc.perms.insert(idx, perm);
+            }
+        }
+
+        if let Some(parent_idx) = parent_idx {
+            self.update_idempotent_foreign_access_after_retag(parent_idx, min_sifa);
+        }
+
+        Ok(())
     }
     fn perform_access(
         &mut self,
@@ -1505,7 +1148,34 @@ impl AllocState for EagerTree {
         alloc_id: AllocId,
         span: Span,
     ) -> UBResult<()> {
-        self.perform_access(tag, access_range, access_kind, access_cause, global, alloc_id, span)
+        #[cfg(feature = "expensive-consistency-checks")]
+        if self.roots.len() > 1 || matches!(prov, ProvenanceExtra::Wildcard) {
+            self.verify_wildcard_consistency(global);
+        }
+
+        let source_idx =
+            if tag.is_wildcard() { None } else { Some(self.tag_mapping.get(&tag).unwrap()) };
+
+        for (loc_range, loc) in self.locations.iter_mut(access_range.start, access_range.size) {
+            let diagnostics = DiagnosticInfo {
+                access_cause,
+                access_range: Some(access_range),
+                alloc_id,
+                span,
+                transition_range: loc_range,
+            };
+            loc.perform_access(
+                self.roots.iter().copied(),
+                &mut self.nodes,
+                source_idx,
+                access_kind,
+                global,
+                ChildrenVisitMode::VisitChildrenOfAccessed,
+                &diagnostics,
+                /* min_exposed_child */ None,
+            )?;
+        }
+        Ok(())
     }
     fn dealloc(
         &mut self,
@@ -1515,7 +1185,69 @@ impl AllocState for EagerTree {
         alloc_id: AllocId,
         span: Span,
     ) -> UBResult<()> {
-        self.dealloc(tag, access_range, global, alloc_id, span)
+        self.perform_access(
+            tag,
+            access_range,
+            AccessKind::Write,
+            AccessCause::Dealloc,
+            global,
+            alloc_id,
+            span,
+        )?;
+
+        let start_idx =
+            if tag.is_wildcard() { None } else { Some(self.tag_mapping.get(&tag).unwrap()) };
+
+        for (loc_range, loc) in self.locations.iter_mut(access_range.start, access_range.size) {
+            let diagnostics = DiagnosticInfo {
+                alloc_id,
+                span,
+                transition_range: loc_range,
+                access_range: Some(access_range),
+                access_cause: AccessCause::Dealloc,
+            };
+            let mut check_tree = |idx| {
+                TreeVisitor { nodes: &mut self.nodes, data: loc }
+                    .traverse_this_parents_children_other(
+                        idx,
+                        |_| ContinueTraversal::Recurse,
+                        |args: NodeAppArgs<'_, _>| {
+                            let node = args.nodes.get(args.idx).unwrap();
+
+                            let perm = args
+                                .data
+                                .perms
+                                .get(args.idx)
+                                .copied()
+                                .unwrap_or_else(|| node.default_location_state());
+                            if global.get_protector_kind(node.tag)
+                                == Some(ProtectorKind::StrongProtector)
+                                && !perm.permission.is_cell()
+                                && perm.accessed
+                            {
+                                Err(TbError {
+                                    error_kind: TransitionError::ProtectedDealloc,
+                                    access_info: &diagnostics,
+                                    conflicting_node_info: &node.debug_info,
+                                    accessed_node_info: start_idx
+                                        .map(|idx| &args.nodes.get(idx).unwrap().debug_info),
+                                }
+                                .build())
+                            } else {
+                                Ok(())
+                            }
+                        },
+                    )
+            };
+            let accessed_root = start_idx.map(&mut check_tree).transpose()?;
+            for &root in self.roots.iter().rev() {
+                if Some(root) == accessed_root {
+                    continue;
+                }
+                check_tree(root)?;
+            }
+        }
+        Ok(())
     }
     fn perform_protector_end_access(
         &mut self,
@@ -1524,12 +1256,73 @@ impl AllocState for EagerTree {
         alloc_id: AllocId,
         span: Span,
     ) -> UBResult<()> {
-        self.perform_protector_end_access(tag, global, alloc_id, span)
+        #[cfg(feature = "expensive-consistency-checks")]
+        if self.roots.len() > 1 {
+            self.verify_wildcard_consistency(global);
+        }
+
+        let source_idx = self.tag_mapping.get(&tag).unwrap();
+
+        let min_exposed_child = if self.roots.len() > 1 {
+            LocationTree::get_min_exposed_child(source_idx, &self.nodes)
+        } else {
+            None
+        };
+
+        for (loc_range, loc) in self.locations.iter_mut_all() {
+            if let Some(p) = loc.perms.get(source_idx)
+                && let Some(access_kind) = p.permission.protector_end_access()
+                && p.accessed
+            {
+                let diagnostics = DiagnosticInfo {
+                    access_cause: AccessCause::FnExit(access_kind),
+                    access_range: None,
+                    alloc_id,
+                    span,
+                    transition_range: loc_range,
+                };
+                loc.perform_access(
+                    self.roots.iter().copied(),
+                    &mut self.nodes,
+                    Some(source_idx),
+                    access_kind,
+                    global,
+                    ChildrenVisitMode::SkipChildrenOfAccessed,
+                    &diagnostics,
+                    min_exposed_child,
+                )?;
+            }
+        }
+        Ok(())
     }
     fn expose_tag(&mut self, tag: BorTag, protected: bool) {
-        self.expose_tag(tag, protected)
+        let id = self.tag_mapping.get(&tag).unwrap();
+        let node = self.nodes.get_mut(id).unwrap();
+        if !node.is_exposed {
+            node.is_exposed = true;
+            let node = self.nodes.get(id).unwrap();
+
+            for (_, loc) in self.locations.iter_mut_all() {
+                let perm = loc
+                    .perms
+                    .get(id)
+                    .map(|p| p.permission())
+                    .unwrap_or_else(|| node.default_location_state().permission());
+
+                let access_level = perm.strongest_allowed_local_access(protected);
+                loc.exposed_cache.update_exposure(
+                    &self.nodes,
+                    id,
+                    WildcardAccessLevel::None,
+                    access_level,
+                );
+            }
+        }
     }
     fn remove_unreachable_tags(&mut self, live_tags: &FxHashSet<BorTag>) {
-        self.remove_unreachable_tags(live_tags)
+        for i in 0..(self.roots.len()) {
+            self.remove_useless_children(self.roots[i], live_tags);
+        }
+        self.locations.merge_adjacent_thorough();
     }
 }
